Move Container_Validator class in the Container namespace/folder
